[FIX] 노트 첫 저장시 존재하지 않는 레코드 오류(스터디 카드셋 삭제) 해결

+++ src/main/java/com/umc/cardify/service/NoteComponentService.java
@@ -142,9 +142,13 @@ public class NoteComponentService {
 	@Transactional
 	public Boolean writeNote(NoteRequest.WriteNoteDto request, Long userId, List<MultipartFile> images) {
 		Note note = noteModuleService.getNoteById(request.getNoteId());
-		cardModuleService.deleteAllCardsByNoteId(note.getNoteId());
-		cardModuleService.deleteAllImageCardsByNoteId(note.getNoteId());
+		System.out.println("note.getNoteId() = " + note.getNoteId());
 
+		if (cardModuleService.existsByNote(note)) {
+			cardModuleService.deleteAllCardsByNoteId(note.getNoteId());
+			cardModuleService.deleteAllImageCardsByNoteId(note.getNoteId());
+
+		}
 		if (!userId.equals(note.getFolder().getUser().getUserId())) {
 			log.warn("Invalid userId: {}", userId);
 			throw new BadRequestException(ErrorResponseStatus.INVALID_USERID);
